third_party/WebKit/LayoutTests/vr/requestPresent_reject_nogesture.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/fake-vr-displays.js"></script>
<script src="resources/mock-vr-service.js"></script>
<canvas id="webgl-canvas"></canvas>
<script src="resources/presentation-setup.js"></script>
<script>
let fakeDisplays = fakeVRDisplays();

vr_test((service) => {
  return navigator.getVRDisplays().then(displays => {
      assert_true(displays != null);
      assert_equals(1, displays.length);
      promise_test( function() {
        return promise_rejects(this, 'InvalidStateError',
            displays[0].requestPresent([{ source : webglCanvas }]));
      }, 'requestPresent rejected');
    });
}, [fakeDisplays['Pixel']],
'Test requestPresent rejects without user gesture');

</script>
